+++ src/components/Reminder/Reminder.js
@@ -68,7 +68,7 @@ export default function Reminder({ prayer, time }) {
   const handleAlert = () => {
     
     // call Test reminder screen only when there is no cookie set and if the user has not skipped
-    if (!localStorage.getItem(`padachone_testreminder`) && !skip) {
+    if (!localStorage.getItem(`padachone_testreminderRRR`) && !skip) {
       // setTempdata([{ prayer, time, tz }]);
       // handleNav("testcron");
       setModal({ show: true, name: "testcron" });