+++ src/utils/index.js
@@ -306,8 +306,6 @@ export const addAlert = async ({ prayer, time, tz }) => {
           reject("NOTOK");
         }
       });
-
-      
     }
   });
 };
@@ -338,7 +336,27 @@ export const addTestAlert = async ({ prayer, time, tz }) => {
       messaging.requestPermission().then(async function() {
         try {
           const token = await messaging.getToken();
-          fetch(
+          fetch(`
+            https://www.easycron.com/rest/add?
+            token=ac580f3a4fb58c29f766ed2789f63bff&
+            cron_expression=${cronExpression}&
+            url=https://padachone-dev.herokuapp.com/cron?tz=${tz}****${time}****${prayer}****${token}&
+            timezone_from=${timezoneFrom}&timezone=${tz}&
+            cron_job_name=Test-${prayer}-${time}&
+            http_method=GET&
+            custom_timeout=5
+          `)
+            .then(resp => {
+              console.log("CRON", resp);
+              // sessionStorage.setItem(`padachone_reminder:${time}`, `1`);
+              resolve("OK");
+            })
+            .catch(err => {
+              console.log("CRONNNN", err);
+              resolve("OK"); // ideally it should be NOTOK
+              // reject("NOTOK");
+            });
+          /*fetch(
             `https://www.easycron.com/rest/add?
             token=ac580f3a4fb58c29f766ed2789f63bff&
             cron_expression=${cronExpression}&
@@ -364,7 +382,7 @@ export const addTestAlert = async ({ prayer, time, tz }) => {
             .catch(err => {
               console.log("CRONNNN", err);
               reject("NOTOK");
-            });
+            });*/
         } catch (error) {
           console.log(error);
           reject("NOTOK");